+++ Makefile
@@ -1,3 +1,4 @@
+
 # A Self-Documenting Makefile: http://marmelab.com/blog/2016/02/29/auto-documented-makefile.html
 
 SOURCE_FILES?=./cmd/mongocli
