+++ internal/cli/atlas/dbusers/delete.go
@@ -12,59 +12,60 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package certs
+package dbusers
 
 import (
 	"github.com/mongodb/mongocli/internal/cli"
 	"github.com/mongodb/mongocli/internal/config"
+	"github.com/mongodb/mongocli/internal/convert"
 	"github.com/mongodb/mongocli/internal/description"
 	"github.com/mongodb/mongocli/internal/flag"
-	"github.com/mongodb/mongocli/internal/json"
 	"github.com/mongodb/mongocli/internal/store"
 	"github.com/mongodb/mongocli/internal/usage"
 	"github.com/spf13/cobra"
 )
 
-type ListOpts struct {
+type DeleteOpts struct {
 	cli.GlobalOpts
-	cli.ListOpts
-	store    store.UserCertificateDescriber
-	username string
+	*cli.DeleteOpts
+	authDB string
+	store  store.DatabaseUserDeleter
 }
 
-func (opts *ListOpts) initStore() error {
+func (opts *DeleteOpts) initStore() error {
 	var err error
 	opts.store, err = store.New(config.Default())
 	return err
 }
 
-func (opts *ListOpts) Run() error {
-	result, err := opts.store.GetUserCertificates(opts.ConfigProjectID(), opts.username)
-
-	if err != nil {
-		return err
-	}
-
-	return json.PrettyPrint(result)
+func (opts *DeleteOpts) Run() error {
+	return opts.Delete(opts.store.DeleteDatabaseUser, opts.authDB, opts.ConfigProjectID())
 }
 
-// mongocli atlas dbuser(s) certs list|ls <username> [--projectId projectId]
-func ListBuilder() *cobra.Command {
-	opts := &ListOpts{}
+// mongocli atlas dbuser(s) delete <username> --force
+func DeleteBuilder() *cobra.Command {
+	opts := &DeleteOpts{
+		DeleteOpts: cli.NewDeleteOpts("DB user '%s' deleted\n", "DB user not deleted"),
+	}
 	cmd := &cobra.Command{
-		Use:     "list <username>",
-		Aliases: []string{"ls"},
-		Short:   description.ListDBUserCerts,
+		Use:     "delete <username>",
+		Aliases: []string{"rm"},
+		Short:   description.DeleteDBUser,
 		Args:    cobra.ExactArgs(1),
 		PreRunE: func(cmd *cobra.Command, args []string) error {
-			opts.username = args[0]
-
-			return opts.PreRunE(opts.initStore)
+			if err := opts.PreRunE(opts.initStore); err != nil {
+				return err
+			}
+			opts.Entry = args[0]
+			return opts.Prompt()
 		},
 		RunE: func(cmd *cobra.Command, args []string) error {
 			return opts.Run()
 		},
 	}
+
+	cmd.Flags().BoolVar(&opts.Confirm, flag.Force, false, usage.Force)
+	cmd.Flags().StringVar(&opts.authDB, flag.AuthDB, convert.AdminDB, usage.AuthDB)
 
 	cmd.Flags().StringVar(&opts.ProjectID, flag.ProjectID, "", usage.ProjectID)
